gcVMOperations.cpp

+++ src/hotspot/share/gc/g1/g1CardSetContainers.hpp
@@ -79,14 +79,14 @@ class G1CardSetInlinePtr : public StackObj {
   static const uintptr_t SizeFieldMask = (((uint)1 << SizeFieldLen) - 1) << SizeFieldPos;
 
   static uint8_t card_pos_for(uint const idx, uint const bits_per_card) {
-    return (idx * bits_per_card + HeaderSize);
+    return narrow_cast<uint8_t>(idx * bits_per_card + HeaderSize);
   }
 
   static ContainerPtr merge(ContainerPtr orig_value, uint card_in_region, uint idx, uint bits_per_card);
 
   static uint card_at(ContainerPtr value, uint const idx, uint const bits_per_card) {
     uint8_t card_pos = card_pos_for(idx, bits_per_card);
-    uint result = ((uintptr_t)value >> card_pos) & (((uintptr_t)1 << bits_per_card) - 1);
+    uint result = narrow_cast<uint>(((uintptr_t)value >> card_pos) & (((uintptr_t)1 << bits_per_card) - 1));
     return result;
   }
